% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{add.rain.info.to.cloud}
\alias{add.rain.info.to.cloud}
\title{Add cloud-top phase information to  summary warm/cold rain mask files into long data frames}
\usage{
add.rain.info.to.cloud()
}
\arguments{
\item{path}{Path to cloud summary files}

\item{years}{Years to process}

\item{out.fname.prefix}{Prefix for output file name (<prefix><years>.rds)}

\item{...}{Optional list of variables to select (lazily evaluated)}
}
\value{
\code{NULL}
}
\description{
Add cloud-top phase information to  summary warm/cold rain mask files into long data frames
}
