% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ceres}
\alias{ceres}
\title{CERES EBAF climatology}
\format{A data frame with the following variables:
\describe{
  \item{toa_sw_all_clim            }{}
  \item{toa_lw_all_clim            }{}
  \item{toa_net_all_clim           }{}
  \item{toa_sw_clr_clim            }{}
  \item{toa_lw_clr_clim            }{}
  \item{toa_net_clr_clim           }{}
  \item{toa_cre_sw_clim            }{}
  \item{toa_cre_lw_clim            }{}
  \item{toa_cre_net_clim           }{}
  \item{solar_clim                 }{}
  \item{cldtau_total_day_clim      }{}
  \item{cldarea_total_daynight_clim}{}
}}
\source{
CERES EBAF Version: Edition 4.0; Release Date March 7, 2017;
DOI: 10.5067/TERRA+AQUA/CERES/EBAF-TOA_L3B.004.0
}
\usage{
ceres
}
\description{
A dataset containing the CERES EBAF 2005-2015 climatology
}
\keyword{datasets}
