% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.cloud}
\alias{combine.cloud}
\title{Combine summary CloudSat cloud/precip flag information summary
files into a long data.frame}
\usage{
combine.cloud(path = "cloud-summaries", years = 2006:2011,
  out.fname.prefix = "cloud-summaries-", vars = c(lon, lat, time, daynight,
  landsea, conv.strat, surface.elev, num.cloud.layers, cloud.flag, precip.flag,
  dardar.precip.flag))
}
\arguments{
\item{path}{Path to cloud summary files}

\item{years}{Years to process}

\item{out.fname.prefix}{Prefix for output file name (<prefix><years>.rds)}

\item{...}{Optional list of variables to select (lazily evaluated)}
}
\value{
\code{NULL}
}
\description{
Combine summary CloudSat cloud/precip flag information summary
files into a long data.frame
}
