% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{df.to.db}
\alias{df.to.db}
\title{Save a data.frame to an SQL database}
\usage{
df.to.db(df, db.name = "wcrain", tbl.name)
}
\arguments{
\item{df}{@code{data.frame} to convert}

\item{fname}{output SQLite db file name}

\item{name}{name of table in db}
}
\value{
produced table as a \code{tbl}
}
\description{
Save a data.frame to an SQL database
}
