% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.cloud.summaries}
\alias{get.cloud.summaries}
\title{Get cloud summary \code{data.frame}}
\usage{
get.cloud.summaries()
}
\description{
This data frame makes it possible to look at more advanced
statistics than just the warm-rain fraction.  For example, one
could look at the probability of warm and cold rain (i.e.,
normalized by the total number of CloudSat footprints) in addition
to the warm-rain fraction (i.e., normalized by the number of
raining CloudSat footprints).
}
