% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.warmfrac_certain}
\alias{get.warmfrac_certain}
\title{Get gridded "rain-certain" warm-rain fraction (2x2 degree lat-lon grid)}
\usage{
get.warmfrac_certain()
}
\description{
Get gridded "rain-certain" warm-rain fraction (2x2 degree lat-lon grid)
}
