% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{grid.warm}
\alias{grid.warm}
\title{Grid warm rain statistics}
\usage{
grid.warm(df = get.cloud.summaries(), bins_lon = 180, bins_lat = 90,
  out.fname = "cloud-warm-grid.rds", precip.classes = 1:3)
}
\arguments{
\item{df}{Data frame to grid, \code{\link{get.cloud.summaries}()} by default}

\item{bins_lon}{Number of bins or vector of bin boundaries (longitude)}

\item{bins_lat}{Number of bins or vector of bin boundaries (latitude)}

\item{out.fname}{File name for output}
}
\description{
Grid warm rain statistics
}
\examples{
## make the \\code{warmrain.rda} file used by \\code{data(\\link{warmrain})}
library(devtools)
warmrain <- grid.warm(get.cloud.summaries(),
                      bins_lon = 180, bins_lat = 90,
                      NULL) 
str(warmrain)
devtools::use_data(warmrain, pkg = "/home/jmuelmen/lib64/R/library/preciptools")


## look at the Great Lakes region, if that's your thing
df <- grid.warm(get.cloud.summaries() \%>\%
               filter(lon > -95, lon < -75, lat > 40, lat < 50),
               seq(-95, -75, 0.1), seq(40, 50, 0.1),
               "great-lakes-grid.rds")
col.frac <- rev(RColorBrewer::brewer.pal(9, "RdYlBu"))[c(1:5,7,9)]
phase.breaks <- c(0, 0.05, 0.1, 0.2 * 1:5)
df \%>\%
    mutate(sum = ice + liq + mixed) \%>\%
    gather(phase, frac, c(ice:mixed, sum)) \%>\%
    discretize(frac, phase.breaks, TRUE) \%>\%
    ggplot(aes(x = lon, y = lat)) +
    geom_raster(aes(fill = frac)) +
    geom_world_polygon(highres = TRUE, lwd = 0.25) +
    xlim(c(-95, -75)) + ylim(c(40,50)) +
    scale_fill_manual(values = col.frac, drop = FALSE) +
    facet_grid(phase ~ denom) +
    coord_fixed(sqrt(2)) +
    theme_bw()
}
