% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label.vertical.features}
\alias{label.vertical.features}
\title{A function to label vertically contiguous features}
\usage{
label.vertical.features(vfm)
}
\description{
A function to label vertically contiguous features
}
