% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.wcrain.R
\name{make.it.rain}
\alias{make.it.rain}
\title{Generate summary warm/cold rain mask files from DARDAR,
2C-PRECIP-COLUMN, and CALMOD}
\usage{

  make.it.rain(path = "ftp.icare.univ-lille1.fr/SPACEBORNE/MULTI_SENSOR/DARDAR_MASK")
}
\arguments{
\item{dardar.path}{Path to DARDAR mask files (used for cloud phase
and DARDAR precip flag)}

\item{2cprecip.path}{Path to 2C-PRECIP-COLUMN files (used for
2C-PRECIP-COLUMN rain flag)}
}
\value{
\code{NULL}
}
\description{
At the end of processing, a summary file containing, for each
CloudSat column, a dry/warm/cold/snow flag based on DARDAR and
2C-PRECIP-COLUMN flags is produced.
}
