% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preciptools.R
\docType{package}
\name{preciptools-package}
\alias{preciptools}
\alias{preciptools-package}
\title{Tools for analyzing precipitation in models and in (satellite) reality}
\description{
Tools for analyzing precipitation in models and in (satellite) reality
}
\author{
\strong{Maintainer}: Johannes Muelmenstaedt \email{johannes.muelmenstaedt@uni-leipzig.de}

}
