% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceres.R
\name{process.ceres}
\alias{process.ceres}
\title{Generate summary data frame from CERES EBAF}
\usage{
process.ceres(infile)
}
\arguments{
\item{infile}{NetCDF file containing CERES EBAF data}
}
\value{
\code{data.frame} containing climatological TOA fluxes
}
\description{
Generate summary data frame from CERES EBAF
}
