% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplemental.R
\name{rain.rate}
\alias{rain.rate}
\title{A supplemental function (supplemental to make.it.rain()) that only
grabs the 2C-PRECIP-COLUMN rain rate}
\usage{

  rain.rate(path = "ftp.icare.univ-lille1.fr/SPACEBORNE/MULTI_SENSOR/DARDAR_MASK")
}
\description{
A supplemental function (supplemental to make.it.rain()) that only
grabs the 2C-PRECIP-COLUMN rain rate
}
