% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{scale_fill_warmfrac}
\alias{scale_fill_warmfrac}
\title{Color scale for warm-rain fraction: \code{\link{discrete_scale}}}
\usage{
scale_fill_warmfrac(name = expression(italic(f)[warm]), labels,
  projector = TRUE)
}
\arguments{
\item{name}{Expression.  Label for the scale}

\item{projector}{Logical.  Use bright color scale appropriate for
projectors (default)}
}
\description{
Color scale for warm-rain fraction: \code{\link{discrete_scale}}
}
