% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{warmfrac.breaks}
\alias{warmfrac.breaks}
\title{Color scale for warm-rain fraction: breaks}
\usage{
warmfrac.breaks(projector = TRUE)
}
\arguments{
\item{projector}{Logical.  Use bright color scale appropriate for
projectors (default)}
}
\description{
Color scale for warm-rain fraction: breaks
}
