% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{warmfrac.col}
\alias{warmfrac.col}
\title{Color scale for warm-rain fraction: color names}
\usage{
warmfrac.col(projector = TRUE)
}
\arguments{
\item{projector}{Logical.  Use bright color scale appropriate for
projectors (default)}
}
\description{
Color scale for warm-rain fraction: color names
}
