% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{warmprob.breaks}
\alias{warmprob.breaks}
\title{Color scale for warm-rain probability: breaks}
\usage{
warmprob.breaks(projector = TRUE)
}
\arguments{
\item{projector}{Logical.  Use bright color scale appropriate for
projectors (default)}
}
\description{
Color scale for warm-rain probability: breaks
}
