% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{warmprob.col}
\alias{warmprob.col}
\title{Color scale for warm-rain probability: color names}
\usage{
warmprob.col(projector = TRUE)
}
\arguments{
\item{projector}{Logical.  Use bright color scale appropriate for
projectors (default)}
}
\description{
Color scale for warm-rain probability: color names
}
