% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{warmrain}
\alias{warmrain}
\title{Gridded warm-rain fractions ("possible drizzle" to "rain certain")}
\format{A data frame with 6 variables:
\describe{
  \item{lon}{longitude; -180 < lon < 180}
  \item{lat}{latitude}
  \item{ice}{fraction of raining ice-topped clouds}
  \item{liq}{fraction of raining liquid clouds}
  \item{mixed}{fraction of raining ice clouds with liquid tops}
  \item{denom}{denominator for fractions (character):
    \describe{
       \item{all}{number of satellite overpasses}
       \item{cloud}{number of cloudy satellite overpasses}
       \item{rain}{number of raining satellite overpasses}
    }}
}}
\source{
See the first example in \code{\link{grid.warm}}, meant to
    be run in \code{/projekt5/climate/jmuelmen} on monsun
}
\usage{
warmrain
}
\description{
A dataset containing the A-Train warm-rain-fraction climatology
}
\keyword{datasets}
